SQL update haha

+++ P2SeriosuGame/SQL/TestParameters.cs
@@ -16,7 +16,7 @@ namespace P2SeriosuGame.SQL
     {
         public int IdTestParameters { get; set; }
         public Nullable<double> Clicks { get; set; }
-        public Nullable<double> Rounds { get; set; }
+        public Nullable<int> Rounds { get; set; }
         public Nullable<int> Wins { get; set; }
         public Nullable<int> Loss { get; set; }
         public Nullable<double> Time_Used { get; set; }
